role/graylog: remove problematic profile::default inclusion
The `profile::default` inclusion is challenging because it manages a
number of unrelated resources. As we switch to using more modules from
the Forge we are likely to see collisions, such as with sudo. At the
moment the `profile::default` class doesn't provide much value so this
commit removes it to avoid the mentioned collisions.