IT-1402: remove hiera hierarchy from lsst-puppet-hiera
As documented in IT-1402 we're incurring a lot of overhead in having
separate Puppet code and Hiera repositories. This commit pairs with
ca6c06de8f in lsst-itconf to move the hiera data out of the
lsst-puppet-hiera hierarchy.

As the lsst-itconf commit changes the hiera hierarchy to ignore this
directory this commit is a noop - but removing this directory will
introduce merge conflicts which will prevent us from committing to the
wrong repository.